backend v2/simps_backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.13.1
gunicorn==20.1.0
jedi==0.17.2
parso==0.7.1
pluggy==1.0.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
ujson==5.2.0
whitenoise==6.0.0
python-environ==0.4.54
